% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch_analysis}
\alias{batch_analysis}
\title{Apply an analysis function to all samples in a dataset}
\usage{
batch_analysis(dataset, fun, cores = 2)
}
\arguments{
\item{dataset}{a metadata table}

\item{fun}{a function that takes a sample and metadata list as first and
second arguments}

\item{cores}{number of cores for parallelization}
}
\value{
a summary data table of analysis results
}
\description{
Applies an analysis function to each sample and binds together
resulting statistics
}
