% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearr_stats.R
\name{compute_deletions}
\alias{compute_deletions}
\title{Compute the distribution of deletion sizes by segment}
\usage{
compute_deletions(sample, metadata = NA)
}
\arguments{
\item{sample}{clonotype table, imported using 'read_mixcr_dataset' or
'read_mixcr_sample' functions}

\item{metadata}{metadata, a named list with chain identifiers
stored under the 'chain' entry - a comma-separated list of chains such as
'TRA', 'TRB', etc is expected. If set to NA will assume all chains are present}
}
\value{
deletion size by segment 2D histogram/table with unweighted 'freq.clonotypes' and
weighted 'freq.reads' frequencies, unnormalized 'count.*' values
are also provided
}
\description{
Computes the histogram of segment trimming sizes for
various segments: 'V3', 'D5', 'D3' or 'J5'.  In case metadata is provided,
segments applicable to computing mean deletion/trimming size are deduced
from the list of chains present in data. Note that values for trimmings
are negative, however, positive values are recorded in case of P-segments
- polyndromic duplications.
}
