% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearr_stats.R
\name{compute_deletions_m}
\alias{compute_deletions_m}
\title{Compute mean deletion sizes by segment}
\usage{
compute_deletions_m(sample, metadata = NA, ...)
}
\arguments{
\item{sample}{clonotype table, imported using 'read_mixcr_dataset' or
'read_mixcr_sample' functions}

\item{metadata}{metadata, a named list with chain identifiers
stored under the 'chain' entry - a comma-separated list of chains such as
'TRA', 'TRB', etc is expected. If set to NA will assume all chains are present}
}
\value{
deletion size by segment table with unweighted 'mean.clonotypes' and
weighted 'mean.reads' means, total number of clonotypes/reads for a given
segment is stored in 'count.*' columns
}
\description{
Computes the means of segment trimming sizes for
various segments: 'V3', 'D5', 'D3' or 'J5'.  In case metadata is provided,
segments applicable to computing mean deletion/trimming size are deduced
from the list of chains present in data. Note that values for trimmings
are negative, however, positive values are recorded in case of P-segments
- polyndromic duplications.
}
