% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearr_stats.R
\name{compute_insertions}
\alias{compute_insertions}
\title{Compute insertion size histogram}
\usage{
compute_insertions(sample, metadata = NA, ...)
}
\arguments{
\item{sample}{clonotype table, imported using 'read_mixcr_dataset' or
'read_mixcr_sample' functions}

\item{metadata}{metadata, a named list with chain identifiers
stored under the 'chain' entry - a comma-separated list of chains such as
'TRA', 'TRB', etc is expected. If set to NA will assume all chains are present}
}
\value{
insert size table with unweighted 'freq.clonotypes' and
weighted 'freq.reads' frequencies, unnormalized 'count.*' values
are also provided
}
\description{
Computes the distribution of insert sizes in various
CDR3 regions: 'VD', 'DJ' or 'VJ'. In case metadata is provided, applicable
insert regions/types are deduced from the list of chains present in data
}
