% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{compute_rearr_stat_dist}
\alias{compute_rearr_stat_dist}
\title{Computes distances between samples in a dataset based on pre-computed
rearrangement statistics}
\usage{
compute_rearr_stat_dist(stats_bundle, value.types = c("reads", "clonotypes"),
  add.pseudocounts = F, cores = 2)
}
\arguments{
\item{stats_bundle}{a set of rearrangement statistics produced by
'compute_rearr_stats' function}

\item{value.types}{specifies whether to use weighted histograms - 'reads' or
to count each clonotype once - 'clonotypes'; can specify both}

\item{add.pseudocounts}{if true will re-compute histograms by adding a pseudocount
of 1 to each histogram, including missing bins, and re-normalize them to new total}

\item{cores}{number of cores for parallelization}
}
\value{
a data frame holding pairwise distances between samples for each
rearrangement statistic
}
\description{
Produces a data frame with pairwise distances between samples computed
for segment usage, insert size and segment trimming size histograms. A square
root of Jensen-Shannon Divergence is used as a metric. Note that no distances
are computed for samples having distinct antigen receptor chain lists
}
