% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{compute_rearr_stat_mds}
\alias{compute_rearr_stat_mds}
\title{Performs a multidimensional scaling placing samples in a 2D plane
according to distance between distributions of rearrangement statistics}
\usage{
compute_rearr_stat_mds(dists_bundle, cores = 2)
}
\arguments{
\item{dists_bundle}{a set of distances produced by
'compute_rearr_stat_dist' function}

\item{cores}{number of cores for parallelization}

\item{value.types}{specifies whether to use weighted histograms - 'reads' or
to count each clonotype once - 'clonotypes'; can specify both}

\item{add.pseudocounts}{if true will re-compute histograms by adding a pseudocount
of 1 to each histogram, including missing bins, and re-normalize them to new total}
}
\value{
a data frame holding sample coordinates for each
rearrangement statistic and chain
}
\description{
Produces a data frame with 2D coordinates for each sample
computed for segment usage, insert size and segment trimming size statistics.
Note that this routine will produce separate placements for each antigen
receptor chain present in dataset
}
