% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{compute_rearr_stats}
\alias{compute_rearr_stats}
\title{Computes rearrangement statistics for a dataset}
\usage{
compute_rearr_stats(dataset, cores = 2)
}
\arguments{
\item{dataset}{a set of samples and metadata produced by 'read_mixcr_dataset'
function}

\item{cores}{number of cores for parallelization}
}
\value{
a list holding 'segment.usage', 'insert.size' and 'deletion.size' tables
}
\description{
Produces a set of data frames holding various rearrangement statistics:
segment usage, insert size histograms, segment trimming histograms
}
