% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearr_stats.R
\name{compute_segment_usage2}
\alias{compute_segment_usage2}
\title{Compute 2D segment usage}
\usage{
compute_segment_usage2(sample, metadata = NA, ...)
}
\arguments{
\item{sample}{clonotype table, imported using 'read_mixcr_dataset' or
'read_mixcr_sample' functions}

\item{metadata}{metadata, a named list with chain identifiers
stored under the 'chain' entry - a comma-separated list of chains such as
'TRA', 'TRB', etc is expected. If set to NA will assume all chains are present}
}
\value{
segment pair usage table with unweighted 'freq.clonotypes' and
weighted 'freq.reads' frequencies, unnormalized 'count.*' values
are also provided
}
\description{
Computes the distribution of segment usage frequencies
for a list of segment type pairs: 'VD', 'DJ' or 'VJ'. By default, all
segments are analyzed. In case metadata is provided, applicable segment
types are deduced from the list of chains present in data
}
