% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{deletion_size_dist}
\alias{deletion_size_dist}
\title{Compute the pairwise Jensen-Shannon distance between samples for segment trimming
size distribution}
\usage{
deletion_size_dist(data, value.types = c("reads", "clonotypes"),
  add.pseudocounts = F, cores = 2, filter.by.chain = F)
}
\arguments{
\item{data}{a set of histograms produced by 'compute_deletions' function}

\item{add.pseudocounts}{if True, will add pseudocounts and re-normalize histograms,
in this case 'value' should contain count, not frequency}

\item{cores}{number of cores for parallelization}

\item{filter.by.chain}{if True, will only compare samples with the same chain,
requires additional 'chain' column in 'data'}
}
\value{
a pairwise distance table, containing sample ids in 'sample.id.1' and
'sample.id.2' columns, statistic type in 'type' column and distance in 'd' column;
if 'filter.by.chain' is set to true will also include 'chain' column
}
\description{
Computes pairwise set of Jensen-Shannon distance provided a set of
per-segment trimming size distribution histograms. If specified, will also only
compare samples with the same antigen receptor chain
}
