% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{dists_to_rdist}
\alias{dists_to_rdist}
\title{Convert pairwise distance tables to dist object}
\usage{
dists_to_rdist(dists)
}
\arguments{
\item{data}{distance table produced by '*_dist' functions}
}
\value{
an R dist object
}
\description{
Converts between distance tables and R dist object
}
