% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{histogram_dist}
\alias{histogram_dist}
\title{Compute the pairwise Jensen-Shannon distance between samples for a set of histograms}
\usage{
histogram_dist(data, cores, add.pseudocounts, filter.by.chain)
}
\arguments{
\item{data}{a set of histograms, should contain 'sample.id', statistic type - 'type',
bin identifier - 'variable' and frequency/count - 'value' columns}

\item{cores}{number of cores for parallelization}

\item{add.pseudocounts}{if True, will add pseudocounts and re-normalize histograms,
in this case 'value' should contain count, not frequency}

\item{filter.by.chain}{if True, will only compare samples with the same chain,
requires additional 'chain' column in 'data'}
}
\value{
a pairwise distance table, containing sample ids in 'sample.id.1' and
'sample.id.2' columns, statistic type in 'type' column and distance in 'd' column;
if 'filter.by.chain' is set to true will also include 'chain' column
}
\description{
Computes pairwise set of Jensen-Shannon distance provided a set of
histograms. Several histogram groups can be provided having different 'type' value.
If specified, will also only compare samples with the same antigen receptor chain
}
