% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mixcr_dataset}
\alias{read_mixcr_dataset}
\title{Read a set of mixcr samples}
\usage{
read_mixcr_dataset(metadata, ...)
}
\arguments{
\item{metadata}{a metadata table}

\item{...}{extra parameters for sample reader}
}
\value{
sample bundle containing metadata and a list of sample data tables
}
\description{
Reads all samples listed in a metadata table containing
'sample.id', 'file.name' and 'chain' columns:
'sample.id' is a unique identifier of a sample,
'file.name' contains full sample path,
'chain' should specify TRA,TRB,IGH,IGK,IGL or their combination
}
