% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mixcr_sample}
\alias{read_mixcr_sample}
\title{Read mixcr sample}
\usage{
read_mixcr_sample(filename, dropExtraColumns = F, verbose = F)
}
\arguments{
\item{filename}{full path to sample}

\item{dropExtraColumns}{drop all columns except ones needed for this package}
}
\value{
sample data table
}
\description{
Reads mixcr clonotype table into a data table
}
